pythonEndpoint/requirements.txt
pyserial
cdataclass